moved AlembicVersion class into alembic_version.py
